// src/types/PlayerStats.ts
export type PlayerStats = {
	name: string
	one: string
	two: string
	three: string
	four: string
	five: string
	six: string
	sum: string
	bonus: string
	onePair: string
	twoPairs: string
	threePairs: string
	threeOfAKind: string
	fourOfAKind: string
	fiveOfAKind: string
	smallStraight: string
	bigStraight: string
	fullStraight: string
	cabin: string
	house: string
	tower: string
	chance: string
	maxiYatzy: string
	totalScore: string
}